[feat] #60 PostEdit 분리
