+++ app/src/main/java/com/dpm/sixpack/main/navigation/MainNavHost.kt
@@ -10,10 +10,11 @@ import com.dpm.sixpack.SixPackAppState
 import com.dpm.sixpack.presentation.destinations.SignInRoute
 import com.dpm.sixpack.presentation.routes.feed.navigation.addFeedNavGraph
 import com.dpm.sixpack.presentation.routes.feed.navigation.navigateToCertifiedUsers
-import com.dpm.sixpack.presentation.routes.feed.navigation.navigateToPostEdit
 import com.dpm.sixpack.presentation.routes.onboarding.navigation.addOnboardingNavGraph
 import com.dpm.sixpack.presentation.routes.postdetail.navigation.addPostDetailNavGraph
 import com.dpm.sixpack.presentation.routes.postdetail.navigation.navigateToPostDetail
+import com.dpm.sixpack.presentation.routes.postedit.navigation.addPostDetailNavGraph
+import com.dpm.sixpack.presentation.routes.postedit.navigation.navigateToPostEdit
 import com.dpm.sixpack.presentation.routes.profilecreation.navigation.addProfileCreationNavGraph
 import com.dpm.sixpack.presentation.routes.profilecreation.navigation.navigateProfileCreation
 import com.dpm.sixpack.presentation.routes.running.navigation.addRunningSessionNavGraph
@@ -130,6 +131,10 @@ internal fun MainNavHost(
                 navigateToUserProfile = {},
                 navigateToMyPage = {},
             )
+
+            addPostDetailNavGraph (
+                navigateToBack = { navController.popBackStack() },
+                )
         }
     }
 }
