+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/feed/navigation/FeedNavigation.kt
@@ -10,7 +10,7 @@ import com.dpm.sixpack.presentation.destinations.MainRoute
 import com.dpm.sixpack.presentation.destinations.PostEdit
 import com.dpm.sixpack.presentation.routes.feed.FeedRoute
 import com.dpm.sixpack.presentation.routes.feed.certifiedusers.CertifiedUsersRoute
-import com.dpm.sixpack.presentation.routes.feed.postedit.PostEditRoute
+import com.dpm.sixpack.presentation.routes.postedit.PostEditRoute
 
 fun NavController.navigateToFeed(navOptions: NavOptions? = null) {
     navigate(MainRoute.Feed, navOptions)
@@ -18,10 +18,6 @@ fun NavController.navigateToFeed(navOptions: NavOptions? = null) {
 
 fun NavController.navigateToCertifiedUsers(date: String) {
     navigate(CertifiedUsers(date = date))
-}
-
-fun NavController.navigateToPostEdit(feedId: Long) {
-    navigate(PostEdit(feedId = feedId))
 }
 
 fun NavGraphBuilder.addFeedNavGraph(
@@ -56,15 +52,6 @@ fun NavGraphBuilder.addFeedNavGraph(
             navigateToBack = navigateToBack,
             navigateToUserProfile = navigateToUserProfile,
             navigateToMyPage = navigateToMyPage,
-        )
-    }
-
-    composable<PostEdit> { backStackEntry ->
-        val route = backStackEntry.toRoute<PostEdit>()
-
-        PostEditRoute(
-            feedId = route.feedId,
-            onNavigateBack = navigateToBack,
         )
     }
 }
